Working refactoring of repository methods
New interface IAuthorRepository
New repository AuthorRepository
New Dependency injection:
builder.Services.AddScoped<AuthorRepository>();
Unit test adapted
